release 0.2
This release comes from continued usage of the containers, which
uncovered some issues that were addressed along the way.
The main focus points were make the images cover all the features
required to reliably run bitbake and improve the reliability of the
build system so that the images build correctly when needed.

This tool has been tested on Fedora 40 & 41. Any feedback on its
operation on other distros will be greatly appreciated.

* add kirkstone and scarthgap releases
* add 32 bit libraries to the images
* improve clarity of some messages
* improve robustness of the detection of images
* do not use the cache when building the images
* pass group into the container and follow the same group as the user
  running on the host
* expose .ssh to the container
* expose bash history to the container

Signed-off-by: António Oliveira <antonio.oliveira@konsulko.com>